Add first prototype of svg elements
Target:
easily operate with svg elements using some unified interface
which takes care of all necessary calculations we have to do before
rendering the svg. In particular scaling of the objects which has
to be done independently along x and y axis and the inversion of
axis. (flip along horizontal axis)